### 2021-09-30 4ed5d3d
Enable strict type checking
