+++ lib/binary_parser.ts
@@ -5,30 +5,27 @@ class Context {
   scopes = [["vars"]];
   bitFields: Parser[] = [];
   tmpVariableCount = 0;
-  references: { [key: string]: { resolved: boolean; requested: boolean } } = {};
+  references = new Map<string, { resolved: boolean; requested: boolean }>();
   importPath: string;
   imports: any[] = [];
   reverseImports = new Map<any, number>();
-  useContextVariables: boolean = false;
+  useContextVariables = false;
 
-  constructor(importPath?: string, useContextVariables?: boolean) {
+  constructor(importPath: string, useContextVariables: boolean) {
     this.importPath = importPath;
     this.useContextVariables = useContextVariables;
   }
 
-  generateVariable(name?: string) {
-    const arr = [];
-
-    const scopes = this.scopes[this.scopes.length - 1];
-    arr.push(...scopes);
+  generateVariable(name?: string): string {
+    const scopes = [...this.scopes[this.scopes.length - 1]];
     if (name) {
-      arr.push(name);
+      scopes.push(name);
     }
 
-    return arr.join(".");
+    return scopes.join(".");
   }
 
-  generateOption(val: number | string | Function) {
+  generateOption(val: number | string | Function): string {
     switch (typeof val) {
       case "number":
         return val.toString();
@@ -40,10 +37,10 @@ class Context {
   }
 
   generateError(err: string) {
-    this.pushCode("throw new Error(" + err + ");");
+    this.pushCode(`throw new Error(${err});`);
   }
 
-  generateTmpVariable() {
+  generateTmpVariable(): string {
     return "$tmp" + this.tmpVariableCount++;
   }
 
@@ -71,7 +68,7 @@ class Context {
     this.scopes.pop();
   }
 
-  addImport(im: any) {
+  addImport(im: any): string {
     if (!this.importPath) return `(${im})`;
     let id = this.reverseImports.get(im);
     if (!id) {
@@ -82,29 +79,37 @@ class Context {
   }
 
   addReference(alias: string) {
-    if (this.references[alias]) return;
-    this.references[alias] = { resolved: false, requested: false };
+    if (!this.references.has(alias)) {
+      this.references.set(alias, { resolved: false, requested: false });
+    }
   }
 
   markResolved(alias: string) {
-    this.references[alias].resolved = true;
+    const reference = this.references.get(alias);
+
+    if (reference) {
+      reference.resolved = true;
+    }
   }
 
   markRequested(aliasList: string[]) {
     aliasList.forEach((alias) => {
-      this.references[alias].requested = true;
+      const reference = this.references.get(alias);
+
+      if (reference) {
+        reference.requested = true;
+      }
     });
   }
 
-  getUnresolvedReferences() {
-    const references = this.references;
-    return Object.keys(this.references).filter(
-      (alias) => !references[alias].resolved && !references[alias].requested
-    );
+  getUnresolvedReferences(): string[] {
+    return Array.from(this.references)
+      .filter(([_, reference]) => !reference.resolved && !reference.requested)
+      .map(([alias, _]) => alias);
   }
 }
 
-const aliasRegistry: { [key: string]: Parser } = {};
+const aliasRegistry = new Map<string, Parser>();
 const FUNCTION_PREFIX = "___parser_";
 const FUNCTION_ENCODE_PREFIX = "___encoder_";
 
@@ -128,7 +133,7 @@ interface ParserOptions {
   trim?: boolean;
   padding?: string;
   padd?: string;
-  tag?: string;
+  tag?: string | ((item: any) => number);
   offset?: number | string | ((item: any) => number);
   wrapper?: (buffer: Buffer) => Buffer;
 }
@@ -316,14 +321,14 @@ export class Parser {
   varName = "";
   type: Types = "";
[…]
