+++ lib/binary_parser.ts
@@ -435,7 +435,7 @@ export class Parser {
 
   private bigIntVersionCheck() {
     if (!DataView.prototype.getBigInt64)
-      throw new Error("BigInt64 is unsupported in this runtime");
+      throw new Error("BigInt64 is unsupported on this runtime");
   }
 
   int64(varName: string, options: ParserOptions = {}): this {
@@ -638,19 +638,22 @@ export class Parser {
   string(varName: string, options: ParserOptions): this {
     if (!options.zeroTerminated && !options.length && !options.greedy) {
       throw new Error(
-        "Neither length, zeroTerminated, nor greedy is defined for string."
+        "One of length, zeroTerminated, or greedy must be defined for string."
       );
     }
+
     if ((options.zeroTerminated || options.length) && options.greedy) {
       throw new Error(
         "greedy is mutually exclusive with length and zeroTerminated for string."
       );
     }
+
     if (options.stripNull && !(options.length || options.greedy)) {
       throw new Error(
-        "Length or greedy must be defined if stripNull is defined."
+        "length or greedy must be defined if stripNull is enabled."
       );
     }
+
     options.encoding = options.encoding || "utf8";
 
     return this.setNextParser("string", varName, options);
@@ -658,7 +661,7 @@ export class Parser {
 
   buffer(varName: string, options: ParserOptions): this {
     if (!options.length && !options.readUntil) {
-      throw new Error("Length nor readUntil is defined in buffer parser");
+      throw new Error("length or readUntil must be defined for buffer.");
     }
 
     return this.setNextParser("buffer", varName, options);
@@ -666,13 +669,11 @@ export class Parser {
 
   wrapped(varName: string, options: ParserOptions): this {
     if (!options.length && !options.readUntil) {
-      throw new Error("Length nor readUntil is defined in buffer parser");
+      throw new Error("length or readUntil must be defined for wrapped.");
     }
 
     if (!options.wrapper || !options.type) {
-      throw new Error(
-        "Both wrapper and type must be defined in wrapper parser"
-      );
+      throw new Error("Both wrapper and type must be defined for wrapped.");
     }
 
     return this.setNextParser("wrapper", varName, options);
@@ -680,19 +681,21 @@ export class Parser {
 
   array(varName: string, options: ParserOptions): this {
     if (!options.readUntil && !options.length && !options.lengthInBytes) {
-      throw new Error("Length option of array is not defined.");
+      throw new Error(
+        "One of readUntil, length and lengthInBytes must be defined for array."
+      );
     }
+
     if (!options.type) {
-      throw new Error("Type option of array is not defined.");
+      throw new Error("type is required for array.");
     }
+
     if (
       typeof options.type === "string" &&
       !aliasRegistry.has(options.type) &&
-      Object.keys(PRIMITIVE_SIZES).indexOf(options.type) < 0
+      !(options.type in PRIMITIVE_SIZES)
     ) {
-      throw new Error(
-        `Specified primitive type "${options.type}" is not supported.`
-      );
+      throw new Error(`Array element type "${options.type}" is unkown.`);
     }
 
     return this.setNextParser("array", varName, options);
@@ -705,13 +708,15 @@ export class Parser {
     }
 
     if (!options) {
-      throw new Error("Option of choice is not defined.");
+      throw new Error("tag and choices are are required for choice.");
     }
+
     if (!options.tag) {
-      throw new Error("Tag option of choice is not defined.");
+      throw new Error("tag is requird for choice.");
     }
+
     if (!options.choices) {
-      throw new Error("Choices option of choice is not defined.");
+      throw new Error("choices is required for choice.");
     }
 
     for (const keyString in options.choices) {
[…]
