+++ test/primitive_parser.ts
@@ -1,25 +1,27 @@
-const assert = require("assert");
-const Parser = require("../dist/binary_parser").Parser;
+import { deepStrictEqual, notDeepStrictEqual, throws, ok } from "assert";
+import { Parser } from "../lib/binary_parser";
 
-const suite = (Buffer) =>
-  describe(`Primitive parser (${Buffer.name})`, function () {
-    function hexToBuf(hex) {
-      return Buffer.from(
-        hex.match(/.{1,2}/g).map((byte) => parseInt(byte, 16))
-      );
+function primitiveParserTests(
+  name: string,
+  factory: (array: Uint8Array | number[]) => Uint8Array
+) {
+  describe(`Primitive parser (${name})`, () => {
+    function hexToBuf(hex: string): Uint8Array {
+      return factory(hex.match(/.{1,2}/g)!.map((byte) => parseInt(byte, 16)));
     }
-    describe("Primitive parsers", function () {
-      it("should nothing", function () {
-        const parser = Parser.start();
 
-        const buffer = Buffer.from([0xa, 0x14, 0x1e, 0x28, 0x32]);
-        assert.deepStrictEqual(parser.parse(buffer), {});
+    describe("Primitive parsers", () => {
+      it("should nothing", () => {
+        const parser = Parser.start();
+        const buffer = factory([0xa, 0x14, 0x1e, 0x28, 0x32]);
+
+        deepStrictEqual(parser.parse(buffer), {});
       });
-      it("should parse integer types", function () {
+      it("should parse integer types", () => {
         const parser = Parser.start().uint8("a").int16le("b").uint32be("c");
 
-        const buffer = Buffer.from([0x00, 0xd2, 0x04, 0x00, 0xbc, 0x61, 0x4e]);
-        assert.deepStrictEqual(parser.parse(buffer), {
+        const buffer = factory([0x00, 0xd2, 0x04, 0x00, 0xbc, 0x61, 0x4e]);
+        deepStrictEqual(parser.parse(buffer), {
           a: 0,
           b: 1234,
           c: 12345678,
@@ -31,11 +33,11 @@ const suite = (Buffer) =>
           it("should parse biguints64", () => {
             const parser = Parser.start().uint64be("a").uint64le("b");
             // from https://nodejs.org/api/buffer.html#buffer_buf_readbiguint64le_offset
-            const buf = Buffer.from([
+            const buf = factory([
               0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x00,
               0x00, 0xff, 0xff, 0xff, 0xff,
             ]);
-            assert.deepStrictEqual(parser.parse(buf), {
+            deepStrictEqual(parser.parse(buf), {
               a: BigInt("4294967295"),
               b: BigInt("18446744069414584320"),
             });
@@ -48,12 +50,12 @@ const suite = (Buffer) =>
               .int64be("c")
               .int64le("d");
             // from https://nodejs.org/api/buffer.html#buffer_buf_readbiguint64le_offset
-            const buf = Buffer.from([
+            const buf = factory([
               0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff, 0x01, 0x00, 0x00,
               0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x00, 0x00, 0xff, 0xff,
               0xff, 0xff, 0x01, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff,
             ]);
-            assert.deepStrictEqual(parser.parse(buf), {
+            deepStrictEqual(parser.parse(buf), {
               a: BigInt("4294967295"),
               b: BigInt("-4294967295"),
[…]
