+++ test/composite_parser.ts
@@ -1,42 +1,42 @@
-const assert = require("assert");
-const Parser = require("../dist/binary_parser").Parser;
+import { deepStrictEqual, throws, doesNotThrow, ok } from "assert";
+import { deflateRawSync, inflateRawSync } from "zlib";
+import { Parser } from "../lib/binary_parser";
 
-const zlib = require("zlib");
-
-const suite = (Buffer) =>
-  describe(`Composite parser (${Buffer.name})`, function () {
-    function hexToBuf(hex) {
-      return Buffer.from(
-        hex.match(/.{1,2}/g).map((byte) => parseInt(byte, 16))
-      );
+function compositeParserTests(
+  name: string,
+  factory: (array: Uint8Array | number[]) => Uint8Array
+) {
+  describe(`Composite parser (${name})`, () => {
+    function hexToBuf(hex: string): Uint8Array {
+      return factory(hex.match(/.{1,2}/g)!.map((byte) => parseInt(byte, 16)));
     }
 
-    describe("Array parser", function () {
-      it("should parse array of primitive types", function () {
+    describe("Array parser", () => {
+      it("should parse array of primitive types", () => {
         const parser = Parser.start().uint8("length").array("message", {
           length: "length",
           type: "uint8",
         });
 
-        const buffer = Buffer.from([12, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]);
-        assert.deepStrictEqual(parser.parse(buffer), {
+        const buffer = factory([12, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]);
+        deepStrictEqual(parser.parse(buffer), {
           length: 12,
           message: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
         });
       });
-      it("should parse array of primitive types with lengthInBytes", function () {
+      it("should parse array of primitive types with lengthInBytes", () => {
         const parser = Parser.start().uint8("length").array("message", {
           lengthInBytes: "length",
           type: "uint8",
         });
 
-        const buffer = Buffer.from([12, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]);
-        assert.deepStrictEqual(parser.parse(buffer), {
+        const buffer = factory([12, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]);
+        deepStrictEqual(parser.parse(buffer), {
           length: 12,
           message: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
         });
       });
-      it("should parse array of user defined types", function () {
+      it("should parse array of user defined types", () => {
         const elementParser = new Parser().uint8("key").int16le("value");
 
         const parser = Parser.start().uint16le("length").array("message", {
@@ -44,10 +44,10 @@ const suite = (Buffer) =>
           type: elementParser,
         });
 
-        const buffer = Buffer.from([
+        const buffer = factory([
           0x02, 0x00, 0xca, 0xd2, 0x04, 0xbe, 0xd3, 0x04,
         ]);
-        assert.deepStrictEqual(parser.parse(buffer), {
+        deepStrictEqual(parser.parse(buffer), {
           length: 0x02,
           message: [
             { key: 0xca, value: 1234 },
@@ -55,10 +55,10 @@ const suite = (Buffer) =>
           ],
         });
       });
-      it("should parse array of user defined types and have access to parent context", function () {
+      it("should parse array of user defined types and have access to parent context", () => {
         const elementParser = new Parser().uint8("key").array("value", {
           type: "uint8",
-          length: function () {
+          length: function (this: any) {
             return this.$parent.valueLength;
           },
         });
[…]
