### 2021-10-04 2bc412b
Translate examples to TypeScript
