### 2021-10-04 b3a6eaa
Make examples work from any directory
