+++ example/mbr.ts
@@ -1,6 +1,8 @@
-import { Parser } from "../lib/binary_parser";
 import { readFile } from "fs";
+import { join } from "path";
 import { inspect } from "util";
+
+import { Parser } from "../lib/binary_parser";
 
 const chs = new Parser()
   .uint8("head")
@@ -39,6 +41,6 @@ const mbrParser = new Parser()
     assert: 0x55aa,
   });
 
-readFile("raspbian.img", function (_, data) {
+readFile(join(__dirname, "raspbian.img"), (_, data) => {
   console.log(inspect(mbrParser.parse(data), { depth: null, colors: true }));
 });
