### 2021-10-04 1adffdd
Make examples work from any directory
