+++ example/bmp.ts
@@ -1,5 +1,7 @@
-import { Parser } from "../lib/binary_parser";
 import { readFile } from "fs";
+import { join } from "path";
+
+import { Parser } from "../lib/binary_parser";
 
 // C structure BITMAPFILEHEADER
 // typedef struct tagBITMAPFILEHEADER {
@@ -56,6 +58,6 @@ const bmpFile = new Parser()
     type: bmpInfoHeader,
   });
 
-readFile("test.bmp", (_, data) => {
+readFile(join(__dirname, "test.bmp"), (_, data) => {
   console.log(bmpFile.parse(data));
 });
