### 2021-10-12 866329d
Run style check with prettier on CI
