+++ .github/workflows/browser.yml
@@ -14,6 +14,6 @@ jobs:
     steps:
     - uses: actions/checkout@v2
     - run: npm ci
-    - run: npm run test-browser
+    - run: npm run test:browser
       env:
         CI: true
