### 2021-10-12 97116ee
Run style check with prettier on CI
