+++ .github/workflows/nodejs.yml
@@ -22,6 +22,7 @@ jobs:
       with:
         node-version: ${{ matrix.node-version }}
     - run: npm ci
+    - run: npm run format
     - run: npm run build --if-present
     - run: npm test
       env:
