### 2021-10-27 7e837fd
Add documentation for sizeOf() in README
