### 2021-11-19 #3777 @majensen
@GaurangTandon would you be available to review [this work](#3777) for [JOSS](https://joss.theoj.org)? We could really benefit from your expertise. Please have a look at the [reviewer information](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Thanks for your consideration! -@majensen
