### 2021-11-19 #3777 @majensen
@Milantex - would you have time and interest to review [this work](#3777) for [https://joss.theoj.org](JOSS)? We would really appreciate your expertise. Please have a look at the [reviewer information](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) and let me know in this thread. Thanks for your consideration! -@majensen
