### 2021-11-20 #3777 @majensen
@Fil - would you be able to review [this work](#3777) for [JOSS](https://joss.theoj.org)? Your JS expertise would be very helpful. Thanks for your consideration - -@majensen
