### 2021-11-21 #3940 @whedon
[REVIEW]: Binary-parser: A declarative and efficient parser generator for binary data
**Submitting author:** @keichi (<a href="http://orcid.org/0000-0002-1607-5694">Keichi Takahashi</a>)
**Repository:** <a href="https://github.com/keichi/binary-parser" target ="_blank">https://github.com/keichi/binary-parser</a>
**Version:** v2.0.1
