### 2021-11-23 d31536b
Fix assert as method definition
