+++ lib/binary_parser.ts
@@ -1011,7 +1011,9 @@ export class Parser {
         );
     }
     ctx.generateError(
-      `"Assertion error: ${varName} is " + ${this.options.assert}`
+      `"Assertion error: ${varName} is " + ${JSON.stringify(
+        this.options.assert.toString()
+      )}`
     );
     ctx.pushCode("}");
   }
