### 2021-11-23 b4080d7
Fix assert as method definition
