### 2021-11-30 #3940 @GaurangTandon
Hi @keichi, I am still going through this project one step at a time. It's a great project! Though, I wanted to discuss regarding "State of the field: Do the authors describe how this software compares to other commonly-used packages" in the software paper. I think your "Design" section is covering both design and comparison to other projects. Is my understanding correct?
