### 2021-12-08 #3940 @Fil
The software works as advertised, and is certainly useful. To complete my review I need help to find the _paper_, _community guidelines_ and _code of conduct_, and more details to see the connection to research purposes (an example of usage in data science would be helpful in that regard). Also I am surprised with https://github.com/keichi/binary-parser/issues/149 and would like to know what the plans are for the governance of this module. thank you! @keichi
