### 2021-12-18 #3940 @GaurangTandon
I reviewed the functionality in this [simple repl](https://replit.com/@SigmaG/OrdinaryWellinformedInstitute#index.js), it works as expected.

The software makes performance claims as well (specifically that the parser generated by this package will perform at least as good as hand written parsers, and sometimes better too). To quote:

> Binary-parser dynamically generates and compiles the parser code on-the-fly, which runs as fast as a hand-written parser (which takes much more time and effort to write).

Unfortunately, I have not been able to test these claims. There is a [benchmarking script](https://github.com/keichi/binary-parser/blob/master/benchmark/bench.js) but it is _just_ a points parser. For a software that makes such a broad performance claim, I'd expect to see - in the repository - at least a dozen different tests to firstly validate this claim.

@Fil You can find the paper link in [this comment](https://github.com/openjournals/joss-reviews/issues/3940#issuecomment-974897731) by the bot. I am not sure about the community guidelines or CoC.
