### 2021-12-18 #3940 @Fil
Thank you @GaurangTandon. For me the last element that needs to be addressed is the community guidelines, in particular wrt issue https://github.com/keichi/binary-parser/issues/149 ; @keichi it would be nice to have some assurance that the software is not going to be abandoned due to a lack of maintainers.
