### 2021-12-28 #3940 @majensen
@keichi - can you please address the points made by @Fil and @GaurangTandon in comments above? In particular,
* Performance claims: either explain the basis of the claims or temper the claims made in the MS
* Create a CONTRIBUTING.md or other means of addressing community guidelines in the repo
* Highlight maybe more clearly the contrast between the state of the field and the contribution that binary-parser makes
thanks and happy new year -
--majensen
