### 2022-04-08 ca50ae7
Escape assert strings
This fixes .string("var", "string with \n in it")
