### 2022-04-08 f682f74
Escape assert strings
This fixes .string("var", "string with \n in it")
