### 2022-05-18 0b4bc49
Allow long bit field sequences
