### 2022-05-18 11e1a49
Allow long bit field sequences
